requirements.txt
numpy
pandas
scikit-learn
flask
matplotlib
seaborn
mysql-connector-python
sqlite3